SMP over BR: Clear LE link key info for temporary bond
In case a temporary bond triggered SMP over BR, the link key would be
dropped in btif layer but btm only reset the classic link key info on
disconnected.

The LE link key info can't be reset together with classic because at
that point we can't distinguish a LTK is generated by the current
temporary bond or not (it may be generated by a previous bond).

As a result, this patch clears the LE flags on SMP_COMPLT_EVT when it's
a temporary bond and there is no LE ACL handle.

Bug: 298459880
Test: mmm packages/modules/Bluetooth
Test: bluetooth_AdapterCLHealth.all_floss
Test: bluetooth_AdapterLEHealth.all_floss
Test: Manually tested NearbyShare and CrossDevice on ChromeOS
Google: 2745255
Change-Id: Ib588fc62dff7aa37ff278ce0657ea14400215760